% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clubs.R
\name{ra_get_club}
\alias{ra_get_club}
\title{Get information about a club}
\usage{
ra_get_club(club_id)
}
\arguments{
\item{club_id}{numeric; club ID (taken from club URL)}
}
\value{
a list
}
\description{
Get information about a club
}
\examples{
\dontrun{
ra_get_club(club_id = 106730)
ra_get_club(club_id = 175569)
}
}
