% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_club_events}
\alias{ra_get_club_events}
\title{Get events for a club}
\usage{
ra_get_club_events(club_id)
}
\arguments{
\item{club_id}{numeric; club ID (taken from club URL)}
}
\value{
a list
}
\description{
Get events for a club
}
\examples{
\dontrun{
ra_get_club_events(club_id = 2587)
ra_get_club_events(club_id = 16687)
}
}
