% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clubs.R
\name{ra_get_club_info}
\alias{ra_get_club_info}
\title{Get clubs information for a club page url}
\usage{
ra_get_club_info(club_url)
}
\arguments{
\item{club_url}{taken form previous function or pasted in}
}
\value{
a list
}
\description{
Get clubs information for a club page url
}
\examples{
\dontrun{
ra_get_club_info("https://www.residentadvisor.net/club.aspx?id=106730")
}
}
