% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clubs.R
\name{ra_get_clubs}
\alias{ra_get_clubs}
\title{Get clubs listed on RA for a region}
\usage{
ra_get_clubs(ai)
}
\arguments{
\item{ai}{numeric; limit results to a specific region
(taken from region's URL e.g. 1 for Sydney)}
}
\value{
a list
}
\description{
Get clubs listed on RA for a region
}
\examples{
\dontrun{
ra_get_clubs(ai = 10)
}
}
