% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_event_info}
\alias{ra_get_event_info}
\title{get_event_info
Get various info from an even page url returns list}
\usage{
ra_get_event_info(event_id)
}
\arguments{
\item{event_id}{numeric; event ID (taken from event URL)}
}
\value{
a list
}
\description{
get_event_info
Get various info from an even page url returns list
}
\examples{
ra_get_event_info(event_id = 1422257)
}
