% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clubs.R
\name{ra_get_region_clubs}
\alias{ra_get_region_clubs}
\title{Get clubs listed on RA for a region}
\usage{
ra_get_region_clubs(ai, include_closed = FALSE)
}
\arguments{
\item{ai}{numeric; limit results to a specific region
(taken from region's URL e.g. 1 for Sydney)}

\item{include_closed}{logical; include clubs that are
now closed in the results?}
}
\value{
a list
}
\description{
Get clubs listed on RA for a region
}
\examples{
\dontrun{
ra_get_region_clubs(ai = 13)
ra_get_region_clubs(ai = 13, include_closed = TRUE)
}
}
