% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_region_events}
\alias{ra_get_region_events}
\title{Get event information for a region}
\usage{
ra_get_region_events(
  region_code,
  start_date,
  date_range = c("month", "week", "day")
)
}
\arguments{
\item{region_code}{region code (taken from \code{ra_get_regions})}

\item{start_date}{date of first event(s) (in ISO 8601 format e.g. "2020-10-01")}

\item{date_range}{date range of events to retrieve (defaults to "month")}
}
\value{
a list
}
\description{
Get event information for a region
}
\examples{
\dontrun{
ra_get_region_events(
region_code = "morocco", start_date = "2020-01-01"
)

ra_get_region_events(
region_code = "uk/manchester", start_date = "2020-09-01"
)
}
}
