% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_regions}
\alias{ra_get_regions}
\title{Get event regions}
\usage{
ra_get_regions()
}
\value{
list
}
\description{
Get the countries and associated regions with event listings.
}
\examples{
\dontrun{
ra_get_regions()
}
}
