% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_region_events}
\alias{ra_get_region_events}
\title{Get event information for a region}
\usage{
ra_get_region_events(country, year, month, region = NULL)
}
\arguments{
\item{country}{country name}

\item{year}{year of event}

\item{month}{month of event}

\item{region}{region name}
}
\value{
a list
}
\description{
Get event information for a region
}
\examples{
\dontrun{
ra_get_region_events(
country = "morocco", year = 2020, month = 1
)
ra_get_region_events(
country = "uk", region = "manchester", year = 2020, month = 9
)
}
}
