% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{ra_get_region_events}
\alias{ra_get_region_events}
\title{Get event information for a region}
\usage{
ra_get_region_events(
  country,
  start_date,
  date_range = c("month", "week", "day"),
  region = NULL
)
}
\arguments{
\item{country}{name of country}

\item{start_date}{date of first event(s) (in ISO 8601 format e.g. "2020-10-01")}

\item{date_range}{date range of events to retrieve (defaults to "month")}

\item{region}{region name (optional; NULL by default)}
}
\value{
a list
}
\description{
Get event information for a region
}
\examples{
\dontrun{
ra_get_region_events(
country = "morocco", start_date = "2020-01-01"
)

ra_get_region_events(
country = "uk", region = "manchester", start_date = "2020-09-01"
)
}
}
